requirements.txt
# WingmanMatch Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & Cache
supabase==2.0.2
redis[hiredis]==5.0.1

# AI & ML
anthropic==0.7.8
openai==1.3.0

# Email Services
resend==0.6.0

# Environment & Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP & Async
httpx==0.24.1
aiofiles==23.2.1

# Retry & Circuit Breaker
tenacity==8.2.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
structlog==23.2.0
# Performance Monitoring Dependencies
asyncpg==0.29.0  # PostgreSQL async connection pool
aiohttp==3.9.1   # For Slack webhook notifications
